Add class to experiment with Redis, add VM
From redistest.rb:
An experiment to get to know Redis's capabilities.
Directly use redis to put a job onto a queue (the representation
of which is glossed over -- e.g. string or whatever), simulate
starting a job, and simulate finishing or failing it.

Also adds a Vagrantfile and playbook for initial provisioning
of a VM, which includes Redis.  You're expected, after spooling
that up, to then 'vagrant ssh' as the 'vagrant' user and go to
/vagrant to set up the wrapper application, etc.

I found that it worked to use the Marmotta / Solr / Jetty
instllation from earlier outside of the VM context, as long as I
destroyed and re-created the Engine Cart application.