WIP on set-awareness prerequisite
Work in progress adding set-awareness to OAI Harvester.  This is
prep work for making an ingestion event that does the right thing
for different types of harvesters.  Some harvesters will want to be
run in parallel (separate jobs for separate sets) and some will need
to be run sequentially.  I'm trying to keep the logic for this
particular to each harvester class.